Allow the use LOGBack appenders
Add an optional dependency to the forthcoming plugin that includes
the LOGBack appender to push events to the NATS pub/sub broker.

LOGBack (https://logback.qos.ch/) is a very powerful and versatile
logging system that allows to decouple the generation of events from
the underlying transport method.

A new option "Enable publish of events to LOGBack" will send all the
events to LOGBack as well, so that other ways of communications
can be implemented to offload all the events outside the Jenkins box.

Change-Id: I595b3a1f3460443cf26c1a68d8581a247f520505